Rename message form file and fix imports
