+++ portfolio/src/components/home/module/return_respond.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect } from 'react';
-import MessageFormProps from './massage_form';
+import { useEffect } from 'react';
+import MessageFormProps from './message_form';
 
 
 interface ReplyProps {
