+++ portfolio/src/components/skilltree/SkillTree.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function SkillTree() {
+  return (
+    <div>
+      <h2>Frontend</h2>
+    </div>
+  );
+}
